Update pip logic
