+++ .env
@@ -1,3 +1,2 @@
 WEB_VIEW_URL=https://docs.belive.sg/views/staging-minimize.html
-
-DEFAULT_SHOW_ID=d12a93bb-fe70-47d4-bafd-d712c12fe1f2
+DEFAULT_SHOW_ID=1355b94e-5c16-4018-86d4-1edc72c37118
