+++ lib/main.dart
@@ -2,14 +2,15 @@ import 'package:flutter/material.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:flutter_in_app_pip/pip_material_app.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
-import 'package:lora_web_view/product_detail_screen.dart';
-import 'package:lora_web_view/show_input_screen.dart';
-import 'package:lora_web_view/webview_screen.dart';
+import 'package:lora_web_view/screens/product_detail_screen.dart';
+import 'package:lora_web_view/screens/show_input_screen.dart';
+import 'package:lora_web_view/screens/webview_screen.dart';
 
 Future main() async {
   await dotenv.load(fileName: ".env");
   runApp(const MainApp());
 }
+
 final GlobalKey<NavigatorState> navigatorKey = GlobalKey<NavigatorState>();
 
 class MainApp extends StatelessWidget {
@@ -34,7 +35,6 @@ class MainApp extends StatelessWidget {
       onGenerateRoute: (settings) {
         if (settings.name == WebViewScreen.routeName) {
           final args = settings.arguments as WebViewScreenArguments;
-
           return MaterialPageRoute(
             builder: (context) {
               return WebViewScreen(
@@ -43,11 +43,15 @@ class MainApp extends StatelessWidget {
               );
             },
           );
-        }
-        if (settings.name == 'product_detail') {
+        } else if (settings.name == ProductDetailScreen.routeName) {
+          final args = settings.arguments as ProductDetailScreenArguments;
           return MaterialPageRoute(
             builder: (context) {
-              return const ProductDetailScreen();
+              return ProductDetailScreen(
+                sku: args.sku,
+                title: args.title,
+                description: args.description,
+              );
             },
           );
         }
